+++ src/converters/lintConfigs/rules/ruleConverters.ts
@@ -64,18 +64,6 @@ import { convertCurly } from "./ruleConverters/curly";
 import { convertCyclomaticComplexity } from "./ruleConverters/cyclomatic-complexity";
 import { convertDeprecation } from "./ruleConverters/deprecation";
 import { convertEofline } from "./ruleConverters/eofline";
-import { convertActionHygiene } from "./ruleConverters/eslint-plugin-ngrx/action-hygiene";
-import { convertAvoidDispatchingMultipleActionsSequentially } from "./ruleConverters/eslint-plugin-ngrx/avoid-dispatching-multiple-actions-sequentially";
-import { convertEffectCreatorAndDecorator } from "./ruleConverters/eslint-plugin-ngrx/effect-creator-and-decorator";
-import { convertNoDispatchInEffects } from "./ruleConverters/eslint-plugin-ngrx/no-dispatch-in-effects";
-import { convertNoDuplicateActionTypes } from "./ruleConverters/eslint-plugin-ngrx/no-duplicate-action-types";
-import { convertNoEffectDecorator } from "./ruleConverters/eslint-plugin-ngrx/no-effect-decorator";
-import { convertNoEffectsInProviders } from "./ruleConverters/eslint-plugin-ngrx/no-effects-in-providers";
-import { convertNoMultipleActionsInEffects } from "./ruleConverters/eslint-plugin-ngrx/no-multiple-actions-in-effects";
-import { convertNoReducerInKeyNames } from "./ruleConverters/eslint-plugin-ngrx/no-reducer-in-key-names";
-import { convertNoTypedStore } from "./ruleConverters/eslint-plugin-ngrx/no-typed-store";
-import { convertOnReducerExplicitReturnType } from "./ruleConverters/eslint-plugin-ngrx/on-reducer-explicit-return-type";
-import { convertSelectorForSelect } from "./ruleConverters/eslint-plugin-ngrx/selector-for-select";
 import { convertJsxBanProps } from "./ruleConverters/eslint-plugin-react/jsx-ban-props";
 import { convertJsxBooleanValue } from "./ruleConverters/eslint-plugin-react/jsx-boolean-value";
 import { convertJsxCurlySpacing } from "./ruleConverters/eslint-plugin-react/jsx-curly-spacing";
@@ -186,6 +174,18 @@ import { convertMochaAvoidOnly } from "./ruleConverters/mocha-avoid-only";
 import { convertNewParens } from "./ruleConverters/new-parens";
 import { convertNewlineBeforeReturn } from "./ruleConverters/newline-before-return";
 import { convertNewlinePerChainedCall } from "./ruleConverters/newline-per-chained-call";
+import { convertActionHygiene } from "./ruleConverters/ngrx-tslint-rules/action-hygiene";
+import { convertAvoidDispatchingMultipleActionsSequentially } from "./ruleConverters/ngrx-tslint-rules/avoid-dispatching-multiple-actions-sequentially";
+import { convertEffectCreatorAndDecorator } from "./ruleConverters/ngrx-tslint-rules/effect-creator-and-decorator";
+import { convertNoDispatchInEffects } from "./ruleConverters/ngrx-tslint-rules/no-dispatch-in-effects";
+import { convertNoDuplicateActionTypes } from "./ruleConverters/ngrx-tslint-rules/no-duplicate-action-types";
+import { convertNoEffectDecorator } from "./ruleConverters/ngrx-tslint-rules/no-effect-decorator";
+import { convertNoEffectsInProviders } from "./ruleConverters/ngrx-tslint-rules/no-effects-in-providers";
+import { convertNoMultipleActionsInEffects } from "./ruleConverters/ngrx-tslint-rules/no-multiple-actions-in-effects";
+import { convertNoReducerInKeyNames } from "./ruleConverters/ngrx-tslint-rules/no-reducer-in-key-names";
+import { convertNoTypedStore } from "./ruleConverters/ngrx-tslint-rules/no-typed-store";
+import { convertOnReducerExplicitReturnType } from "./ruleConverters/ngrx-tslint-rules/on-reducer-explicit-return-type";
+import { convertSelectorForSelect } from "./ruleConverters/ngrx-tslint-rules/selector-for-select";
 import { convertNoAngleBracketTypeAssertion } from "./ruleConverters/no-angle-bracket-type-assertion";
 import { convertNoArg } from "./ruleConverters/no-arg";
 import { convertNoAsyncWithoutAwait } from "./ruleConverters/no-async-without-await";
