+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/action-hygiene.ts
