+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/avoid-dispatching-multiple-actions-sequentially.ts
