+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/effect-creator-and-decorator.ts
