+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/no-dispatch-in-effects.ts
