+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/no-duplicate-action-types.ts
