+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/no-effect-decorator.ts
