+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/no-effects-in-providers.ts
