+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/no-multiple-actions-in-effects.ts
