+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/no-multiple-stores.ts
