+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/no-reducer-in-key-names.ts
