+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/no-typed-store.ts
