+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/on-reducer-explicit-return-type.ts
