+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/selector-for-select.ts
