+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/action-hygiene.test.ts
