+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/avoid-dispatching-multiple-actions-sequentially.test.ts
