+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/effect-creator-and-decorator.test.ts
