+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/no-dispatch-in-effects.test.ts
