+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/no-duplicate-action-types.test.ts
