+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/no-effect-decorator.test.ts
