+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/no-effects-in-providers.test.ts
