+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/no-multiple-actions-in-effects.test.ts
