+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/no-multiple-stores.test.ts
