+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/no-reducer-in-key-names.test.ts
