+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/no-typed-store.test.ts
