+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/on-reducer-explicit-return-type.test.ts
