+++ src/converters/lintConfigs/rules/ruleConverters/ngrx-tslint-rules/tests/selector-for-select.test.ts
